chore: componentizing of category and tag table removed
